pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "0.1.0"
description = "Email Triage OpenEnv Environment - Train AI agents to prioritize, categorize, and manage inbox emails"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Email Triage OpenEnv Contributors"}
]
keywords = ["openenv", "reinforcement-learning", "email", "triage", "ai-agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0.0",
    "openenv-core>=0.1.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
server = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "httpx>=0.24.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "scipy>=1.10.0",
]
inference = [
    "openai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "models*", "client*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
